MLK-14363 mx6qsabreauto: Fix ethernet PHY setting issue
The PHY settings for RGMII has been removed from mx6qsabreauto board codes,
due to the atheros PHY driver have updated to use same configuration for
AR8031 and AR8035, while this configuration is duplicated as we set in board codes.

But in recent codes, the PHY driver added a patch for AR8031 independent config.
So needs to add the PHY settings back to the board codes.

Signed-off-by: Ye Li <ye.li@nxp.com>