MLK-15026 usb: host: xhci-imx8: add imx8 USB3 driver
It is for imx8qm and imx8qxp, and the cadence IP is in it.

Both DM and Non-DM are supported in this driver. The DM support
only probes the USB3 node, but won't parse any properties from node in DTS.
The registers address are still hard coded, that share with non-DM codes.

The DM driver will enable the power of USB3 controller and USB3 PHY. But
for Non-DM driver, users need to power up them in board level codes.

Signed-off-by: Peter Chen <peter.chen@nxp.com>
Signed-off-by: Ye Li <ye.li@nxp.com>