MLK-16752 imx8mq: Add CPU rev get and use it for ROM_SW_INFO address
Implement the get_cpu_rev function by reading the DIGPROG register
to get the CPU ID. Since this register is not updated for B0,
we need also check ROM version. The ROM version address is different
on B0 (0x83c) and A0 (0x800), so have to check both two.

Additional, the address for ROM_SW_INFO is changed on B0, we have to use
the CPU version to determine the address.

Signed-off-by: Ye Li <ye.li@nxp.com>