MLK-17313-1 tcpc: Add support for power sink in dead battery
To add support for power sink when booting in dead battery, we
have to setup PD connection with PD source and send/receive PD messages
to negotiate a proper voltage and current.

This patch refactors the TCPC driver with this new function and also
keep DFP mode support for USB host.

Signed-off-by: Ye Li <ye.li@nxp.com>
Acked-by: Jun Li <jun.li@nxp.com>