MLK-16083 imx8qm/qxp: Setup MMU pagetables dynamically with owned memory regions
Since the memory regions assigned to u-boot partition is dynamically set by
ATF and SCD. We have to setup MMU pagetables according to the owned memory
regions, not set it for unassigned memory.

Additional, the memory region 0x0-0x1c000000 is for LSIO subsystem, including OCRAM,
AP ROM, flexspi0 mapped memory and flexspi1 buffer. If we set it to cachable,
the AHB read in flexspi driver will have coherence problem. So we set this memory
region to strongly order to avoid any issue in driver.

Signed-off-by: Ye Li <ye.li@nxp.com>