TEE-330: i.MX8QM: OP-TEE: increase number of DRAM BANKS
When OP-TEE is enabled a share memory is created between Linux and OP-TEE.
On i.MX 8QM with the use of the SC subsystem, this creates a new chunk of
memory.
Add a new DRAM BANK, so that u-boot is able to capture correctly all
the chunks of memory in meminfo.

Signed-off-by: Silvano di Ninno <silvano.dininno@nxp.com>
(cherry picked from commit c89c493cd8657f5d520647c3a9a61861b3f267e0)