MLK-20510-2 mx7d: Align implementation of is_boot_from_usb with mx6
Because the ROM info on iMX7D does not set device type to USB when booting
from serial download mode, we have to use the mechanism on mx6 to implement
the is_boot_from_usb. The original implementation is checking USB controller
register, it can't work correctly after any USB functionality is run in u-boot.

Signed-off-by: Ye Li <ye.li@nxp.com>